Fixes based on the teacher recommendations
