+++ project/src/services/handle-error.ts
@@ -2,12 +2,12 @@ import request from 'axios';
 import { toast } from 'react-toastify';
 
 import { HttpCode } from '../constants/http';
-import { TError } from '../types/error';
+import { Error } from '../types/error';
 import { store } from '../store';
 import { redirectToRoute } from '../store/action';
 import { AppRoute } from '../constants/routs';
 
-export function handleError(error: TError): void {
+export function handleError(error: Error): void {
   if (!request.isAxiosError(error)) {
     throw error;
   }
@@ -17,8 +17,6 @@ export function handleError(error: TError): void {
   if (response) {
     switch (response.status) {
       case HttpCode.BAD_REQUEST:
-        toast.info(response.data.error);
-        break;
       case HttpCode.UNAUTHORIZED:
         toast.info(response.data.error);
         break;
