+++ project/src/store/constants.ts
@@ -1,0 +1,8 @@
+export const CHANGE_GENRE = 'CHANGE_GENRE';
+export const LOAD_FILMS = 'LOAD_FILMS';
+export const LOAD_SIMILAR_FILMS = 'LOAD_SIMILAR_FILMS';
+export const LOAD_FILM = 'LOAD_FILM';
+export const LOAD_PROMO = 'LOAD_PROMO';
+export const REQUIRE_AUTHORIZATION = 'REQUIRE_AUTHORIZATION';
+export const LOAD_COMMENTS = 'LOAD_COMMENTS';
+export const REDIRECT_TO_ROUTE = 'REDIRECT_TO_ROUTE';
