+++ project/src/store/reducer.ts
@@ -9,40 +9,9 @@ import {
   loadSimilarFilms,
   requireAuthorization
 } from './action';
-import { Film } from '../types/film';
 import { GenreEnum } from '../constants/genres';
 import { AuthorizationStatus } from '../constants/auth';
-import { Comment } from '../types/comment';
-
-export type InitialState = {
-  genre: GenreEnum,
-  authorizationStatus: AuthorizationStatus,
-  films: {
-    data: Film[],
-    isLoading: boolean,
-    isLoaded: boolean,
-  },
-  similarFilms: {
-    data: Film[],
-    isLoading: boolean,
-    isLoaded: boolean,
-  },
-  promo: {
-    data: Film | null,
-    isLoading: boolean,
-    isLoaded: boolean,
-  },
-  film: {
-    data: Film | null,
-    isLoading: boolean,
-    isLoaded: boolean,
-  },
-  comments: {
-    data: Comment[],
-    isLoading: boolean,
-    isLoaded: boolean,
-  },
-};
+import { InitialState } from '../types/initial-state';
 
 const initialState: InitialState = {
   genre: GenreEnum.ALL_GENRES,
