+++ project/src/types/error.ts
@@ -1,1 +1,1 @@
-export type TError = unknown;
+export type Error = unknown;
